+++ examples/kyb-app/src/common/providers/ViewStateProvider/components/ViewResolver/types.ts
@@ -1,0 +1,1 @@
+export type Views<TKeys extends PropertyKey> = Record<TKeys, React.ComponentType<unknown>>;