+++ examples/kyb-app/src/components/organisms/KYBView/hooks/useFlowMutation/useFlowDataMutation.ts
@@ -1,0 +1,7 @@
+import { updateFlowData } from '@app/domains/workflows';
+
+export const useFlowDataMutation = () => {
+  return {
+    mutate: updateFlowData,
+  };
+};