+++ services/workflows-service/src/business/business.controller.external.ts
@@ -17,6 +17,7 @@ import { WorkflowDefinitionModel } from '@/workflow/workflow-definition.model';
 import { WorkflowDefinitionFindManyArgs } from '@/workflow/dtos/workflow-definition-find-many-args';
 import { WorkflowService } from '@/workflow/workflow.service';
 import { makeFullWorkflow } from '@/workflow/utils/make-full-workflow';
+import { BusinessUpdateDto } from '@/business/dtos/business.update';
 
 @swagger.ApiTags('external/businesses')
 @common.Controller('external/businesses')
@@ -78,6 +79,24 @@ export class BusinessControllerExternal {
 
       throw err;
     }
+  }
+
+  @common.Put(':id')
+  @UseKeyAuthInDevGuard()
+  async update(@common.Param('id') businessId: string, @common.Body() data: BusinessUpdateDto) {
+    return this.service.updateById(businessId, {
+      data: {
+        companyName: data.companyName,
+        address: data.address,
+        registrationNumber: data.registrationNumber,
+        website: data.website,
+        documents: data.documents ? JSON.stringify(data.documents) : undefined,
+        shareholderStructure:
+          data.shareholderStructure && data.shareholderStructure.length
+            ? JSON.stringify(data.shareholderStructure)
+            : undefined,
+      },
+    });
   }
 
   // curl -v http://localhost:3000/api/v1/external/businesses/:businessId/workflows